disrete: update ordering in docstring

+++ gymnasium/spaces/discrete.py
@@ -50,7 +50,7 @@ class Discrete(Space[IntType]):
 
         Args:
             n (int): The number of elements of this space.
-            dtype: The space type, for example, ``int``, ``np.int64``, ``np.int32``, or ``np.uint8``.
+            dtype: The space type, for example, ``int``, ``np.uint8``, ``np.int32``, or ``np.uint8``.
             seed: Optionally, you can use this argument to seed the RNG that is used to sample from the ``Dict`` space.
             start (int): The smallest element of this space.
         """
